fix: 입력게이트 락 + 그로기 다이아 + 피격 float 캐스팅

+++ NAN2026Game/Assets/Scripts/ThrownProjectile.cs
@@ -71,7 +71,7 @@ namespace NAN2026
             }
             if (dist <= 0.55f)
             {
-                player.SendMessage("TakeDamage", config.damage, SendMessageOptions.DontRequireReceiver);
+                player.SendMessage("TakeDamage", (float)config.damage, SendMessageOptions.DontRequireReceiver);
                 Destroy(gameObject);
             }
         }
